requirements.txt
APScheduler==3.11.0
Django==5.1.4
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
psycopg2-binary
requests==2.32.3


gunicorn
whitenoise
dj-database-url
python-decouple